release commit

+++ src/main/java/ru/mccormick/ipa/dao/CalculationMapper.java
@@ -21,7 +21,7 @@ public class CalculationMapper implements RowMapper<Calculation> {
 		calculation.setCalculationId(rs.getInt("calculation_id"));
 		calculation.setUserId(rs.getInt("user_id"));
 		//calculation.setCalculationValues(calculationValuesService.getValuesList(rs.getInt("calculation_id")));
-		calculation.setCalculationResult(rs.getInt("calculation_result"));
+		calculation.setCalculationResult(rs.getDouble("calculation_result"));
 		
 		return calculation;
 	}
